Time: 102 ms (6.70%), Space: 116.1 MB (53.61%) - LeetHub

+++ 3101-count-alternating-subarrays/3101-count-alternating-subarrays.cpp
@@ -6,7 +6,7 @@ public:
         int n = nums.size();
         while(j<n){
             if(nums[j]==nums[j-1]){
-                int len = j-i;
+                long long len = j-i;
                 long long temp = (len*(len+1))/2;
                 ans+=temp;
                 i = j;
